(refactor) Rename component names for clarity
