+++ client/src/components/containers/VoteBoard.js
@@ -3,11 +3,11 @@ import '../../styles/css/temp.css';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
 import * as actionCreators from '../../actions/actionCreators';
-import NomineesList from '../presentational/Nominations_NomineesList';
+import NomineeList from '../presentational/PollingList';
 import WinningResult from '../presentational/WinningResult';
 import Lobby from '../presentational/Lobby';
 
-class NominationsBoard extends Component {
+class VoteBoard extends Component {
 
   componentWillMount() {
     // fetch commitments based on user
@@ -25,7 +25,7 @@ class NominationsBoard extends Component {
           </div>
           <div>
             {this.props.nominees.map((nominee, i) =>
-              <NomineesList
+              <NomineeList
                 key={i}
                 nominee={nominee}
               />
@@ -59,4 +59,4 @@ function mapDispatchToProps(dispatch) {
   return bindActionCreators(actionCreators, dispatch);
 }
 
-export default connect(mapStateToProps, mapDispatchToProps)(NominationsBoard);
+export default connect(mapStateToProps, mapDispatchToProps)(VoteBoard);
