+++ client/src/components/presentational/Polling_ControlBoard.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import '../../styles/css/temp.css';
-import VoteBoard from '../containers/Nominations_ControlBoard_VoteBoard';
+import VoteBoard from '../containers/Polling_ControlBoard_VoteBoard';
 
 class ControlBoard extends Component {
