+++ client/src/components/presentational/Polling_ControlBoard_UserStatus.js
@@ -1,17 +1,17 @@
 import React, { Component } from 'react';
 import '../../styles/css/temp.css';
 
-class VoteList extends Component {
+class UserStatus extends Component {
 
   render() {
     const { suggestion } = this.props;
     return (
       <div>
         <div>{ suggestion.locationName }</div>
-        <div>Vote Count</div>
+        <div>User Status</div>
       </div>
     );
   }
 };
 
-export default VoteList;
+export default UserStatus;
