+++ client/src/components/presentational/Polling_Page.js
@@ -1,19 +1,19 @@
 import React, { Component } from 'react';
 import '../../styles/css/temp.css';
-import NominationsBoard from '../containers/Nominations_Board';
-import NominationsControlBoard from '../presentational/Nominations_ControlBoard';
+import PollingBoard from '../containers/VoteBoard';
+import PollingControlBoard from '../presentational/Polling_ControlBoard';
 
 //think of this as simply a wrapper for the Nominations page
-class Nominations extends Component {
+class Polling extends Component {
   // This page will contain components for the chat feature when we get around to it
   render() {
     return (
       <div>
-        <NominationsBoard />
-        <NominationsControlBoard />
+        <PollingBoard />
+        <PollingControlBoard />
       </div>
     );
   }
 };
 
-export default Nominations;
+export default Polling;
